requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pandas==2.2.2
openpyxl==3.1.5
networkx==3.3
sqlalchemy==2.0.32
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.32.3
streamlit==1.38.0
plotly==5.24.1
pytest==8.3.3
fastapi
uvicorn
pandas
networkx
plotly
streamlit
requests
scipy
python-dotenv
groq 